[UPDATE] main controller bcryptjs

+++ src/controllers/mainController.js
@@ -1,5 +1,6 @@
 const fs = require("fs");
 const path = require("path");
+const bcryptjs = require('bcryptjs');
 const { validationResult } = require("express-validator");
 
 const usersPath = path.resolve(__dirname, "../database/users.json");
@@ -9,9 +10,11 @@ module.exports = {
   home: (req, res) => {
     res.render("home");
   },
+
   register: (req, res) => {
     res.render("client/register");
   },
+
   createUser: (req, res) => {
     const valResult = validationResult(req);
 
@@ -21,15 +24,30 @@ module.exports = {
         oldData: req.body,
       });
     } else {
+      const { first_name, last_name, email, birth_date, password, confirmation_pwd } = req.body;
+
       let newUser = {
         id: users[users.length - 1].id + 1,
-        ...req.body,
+        first_name,
+        last_name,
+        email,
+        birth_date,
+        password: bcryptjs.hashSync(password, 9)
       };
-      users.push(newUser);
-      fs.writeFileSync(usersPath, JSON.stringify(users, null, " "));
-      res.redirect("/login");
+
+      if (bcryptjs.compareSync(confirmation_pwd, newUser.password)) {
+        users.push(newUser);
+        fs.writeFileSync(usersPath, JSON.stringify(users, null, " "));
+        res.redirect("/login");
+      } else {
+        return res.render("client/register", {
+          // errors: ??? ,
+          oldData: req.body,
+        })
+      }
     }
   },
+
   login: (req, res) => {
     res.render("client/login");
   },
